Move ApiObjects to JsonApi namespace/folder
Follow up to
https://github.com/rails-api/active_model_serializers/pull/1504/files#r52796473.
Moved the Relationship and ResourceIdentifier to the JsonApi
namespace/folder and slightly refactored the Relationship code.
